0.965 較正のやり直しを音と表示で知らせる。音の仕様書を追加
■ ESKF CALIBRATION REQUIRED! が画面だけだった
この状態では姿勢表示・バンク角警告・自動ロールトリムがまとめて信用できないが、
地上では画面を見ていないことが多く、気づかないまま飛べてしまっていた。
・音でも促す（120 秒ごと）。SD が無い機体では 262Hz×3 のトーンで代替。
  ★ 鳴らすのは地上で静止しているときだけ。飛行中に較正はできないので、
    対処できない警告で注意をそらすほうが危険。地上バンクチェックと
    同じブロックに置いて、その入口条件（地上・静止・リプレイ/ミラー以外）を共有する。
・表示は 3 秒のうち 1 秒だけの点滅にした。万一この状態で飛んでしまったとき、
  地図を隠し続けないため。
・ヘッダーの Pavg に赤い横バーを重ねて、その数字が信用できないことを示す。
  位置と太さは「GNSS 0 個で速度・方位が無効」の赤線と揃えた（同じ意味を同じ見た目で）。
wav/eskf_calib_required.wav を追加。

■ 最低保証音量(60)の乱用をやめた
「持ち主が機体から離れていても電池切れに気づけるように」という目的で入れた仕組みだが、
link_alert() が無線の警告 8 種すべてに付けていたため、音量を絞っても
それらが大音量で鳴っていた。link_alert() の min_vol を既定 0 にし、
battery_low_sender だけに 60 を渡す。リプレイ中止のトーンからも外した。
残るのは自機とセンサー機の電池低下だけ。関数のコメントに理由と経緯を残した。

■ Auto10km の折返し通知の鳴り方が往復で違っていた
往路→復路は通常トーン（音声と重なる）、復路→往路は solo_play（音声が後）。
同じ通知なので solo_play に揃えた。入れ替え防止のコメントを両方に置いた。

■ 「トラック」警報のしきい値が実質 16 度だった
update_tone() が方位差を int で受けていたため 15.9 度が 15 に切り捨てられていた。
float にして仕様どおり 15 度で発報する。

■ docs/pons_sound.md を新規作成
WAV / トーン / バリオの 3 系統、Core の分担、優先度、同時再生の可否、音量、
バリオの実測値、音声ファイル 30 件の一覧（優先度と発報条件つき）、
音声を持たない音の一覧、音で知らせていない警告。
README と CLAUDE.md から辿れるようにした。

RELEASE と非 RELEASE の両構成でビルド確認済み。新規の警告なし。

+++ src/sound.cpp
@@ -6,7 +6,7 @@
 //           旋回角速度(degpersecond)に応じた音程変化、
 //           アンプシャットダウン制御（省電力）。
 // Author  : MasaoC (@masao_mobile)
-// Updated : 2026/09/12
+// Updated : 2026/09/14
 // ============================================================
 // Handle speaker, amplifier, PWM-audio signals.
 #include "../settings.h"
@@ -782,12 +782,15 @@ void update_tone(float degpersecond){
     last_update_tone = millis();
 
   // 前回の方位との差を計算し、-180〜+180 の範囲に収める
-  int relativedif = get_gps_truetrack()-last_tone_tt;
+  // ★ float で持つこと。以前は int で受けていたため 15.9 度が 15 に切り捨てられ、
+  //   下の「15 度超」が実際には 16 度超になっていた（音で鳴る警報なので影響が出る）。
+  //   last_tone_tt 側は full precision で latch してあるので、粗かったのは比較だけ。
+  float relativedif = get_gps_truetrack()-last_tone_tt;
   if(relativedif > 180)
     relativedif -= 360;
   if(relativedif < -180)
     relativedif += 360;
-  int angle_diff = abs(relativedif);
+  float angle_diff = abs(relativedif);
 
   //【警告1】方位変化が 15° 超 → コース逸脱警告
   if(angle_diff > 15){
